Install/Uninstall improvements.
Now defining a development safe default for CMAKE_INSTALL_PREFIX.

Removed the UNINSTALL target because there is no way to safely support custom targets over arbitrarily deep nested projects using add_subdirectory. In order to avoid a target name conflict projects would be required to declare the UNINSTALL target before adding any sub-directory and also test if the UNINSTALL target already exists before declaring one. Neither of these requirements can be enforced specially when adding 3rdparty submodules. And there is also the complication of knowing when to define the uninstall target in upper case vs. lower case which may depend on the generator used. So for now let's assume uninstall goes in the same category of clean, that is, it has to be done manually, at least for sub-projects (which all libraries can potentially be by definition). Projects generating executable(s) as their main target may choose otherwise because they tend to be top projects only but even those would have to be considered on a case-by-case basis because multiple executable projects could still be grouped by a top project interested in constructing and installing a collection of tools for example.